hm-06: set clang-tidy in the repo root, fixing warnings

+++ homeworks/06_tests_code_quality/src/ballistics.cpp
@@ -4,20 +4,21 @@
 #include <cstring>
 #include <iostream>
 
-float get_h(float t, float d, float g, float l, float m, float v0)
+float get_h(
+  float t, float d, float g, float l, float m, float v0)  // NOLINT(readability-identifier-length), formula values, easier to handle
 {
-  float l2 = pow(l, 2);
+  float l2 = powf(l, 2);
   float l2p1 = l2 + 1;
 
-  float term1 = (pow(t, 3) * (6 * d * g * l * m - 6 * pow(d, 2) * (l2 - 1) * v0)) / (36 * pow(m, 2));
+  float term1 = (powf(t, 3) * (6 * d * g * l * m - 6 * powf(d, 2) * (l2 - 1) * v0)) / (36 * powf(m, 2));
 
-  float term2 = (pow(t, 5) * (3 * pow(d, 3) * g * pow(l, 3) * m - 3 * pow(d, 4) * l2 * l2p1 * v0)) / (36 * l2p1 * pow(m, 4));
+  float term2 = (powf(t, 5) * (3 * powf(d, 3) * g * powf(l, 3) * m - 3 * powf(d, 4) * l2 * l2p1 * v0)) / (36 * l2p1 * powf(m, 4));
 
-  float term3 = (pow(t, 4) * (3 * pow(d, 3) * (l2p1)*l2 * v0 + 6 * pow(d, 3) * l2p1 * pow(l, 4) * v0 -
-                              6 * pow(d, 2) * g * (pow(l, 4) + l2p1) * l * m)) /
-                (36 * pow(l2p1, 2) * pow(m, 3));
+  float term3 = (powf(t, 4) * (3 * powf(d, 3) * (l2p1)*l2 * v0 + 6 * powf(d, 3) * l2p1 * powf(l, 4) * v0 -
+                               6 * powf(d, 2) * g * (powf(l, 4) + l2p1) * l * m)) /
+                (36 * powf(l2p1, 2) * powf(m, 3));
 
-  float term4 = (d * pow(t, 2) * v0) / (m * 2);
+  float term4 = (d * powf(t, 2) * v0) / (m * 2);
 
   return term1 + term2 + term3 - term4 + (t * v0);
 }
